requirements.txt
praw
google-generativeai
Pillow
python-dotenv
streamlit